Callbacks and debug work when capture is enabled (#1902)
**Context:**

As we want to make capture the new frontend for catalyst, we need to
make sure all catalyst features continue to work.

**Description of the Change:**

Update the `callback` and `debug` code so that it still works with
capture enabled.

**Benefits:**

More functionality!

**Possible Drawbacks:**

I'm needing to leave the primitives and frontend in catalyst, as the
code can't really be separated from it's lower level components, and I'm
trying to be time efficient.

I also was seeing problems trying to get
`test_debug.py:TestCProgramGeneration` and
`test_debug.py:TestOptionsToCliFlags` to pass with both frontends, so I
decided not to add capture tests for them in this PR. We may need to
come back to investigating those tests.

**Related GitHub Issues:**

[sc-95499]