Add dialect docs infra and static generated outputs (#1959)
**Context:** To ensure we can have accessible MLIR dialect and pass
docs, we require adding the generated output of the `mlir-doc` target to
the site documentation. As this requires a full clone of catalyst and
all submodules, we add the files statically based on the latest
generated output.

**Description of the Change:** Add dialect and pass docs to the sphinx
site as static additions. Add infra changes to allow easy generation of
the dialect and pass documentation.

**Benefits:** Ensures the above are easily accessible.

**Possible Drawbacks:** Static generated files may grow stale, and will
require periodic updates.

**Related GitHub Issues:**

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>