Issue 1220 add merge rotation patterns for qml rot and qml crot (#1955)
**Context:**

This task has to do with Catalyst's Merge Rotation pass.
When this pass was originally implemented, arbitary rotations were part
of the list of rotations that could be merged.
Only to find later on that, for two arbitrary rotations, the merge was
not as easy as for a fixed rotation, e.g., `RX`, where the rotation
angles are simply added. At that point, arbitary rotations were removed
from the set of mergeable rotations.
For an arbitary rotation, more complex mathematical computations have to
be performed in order to obtain the angles of the merged gate. These
mathematical formulas are well described
[here](https://docs.pennylane.ai/en/stable/code/api/pennylane.transforms.single_qubit_fusion.html#derivation).
In order to add arbitrary rotations back, issue #1220 was created. This
issue is also very well documented.
This PR tries to address that issue.

**Description of the Change:**

1. Add arbitrary rotations, `Rot` and `CRot` to the list of rotations in
MergeRotationsPatterns.cpp.
2. Change the logic of `MergeRotationsRewritePattern::matchAndRewrite`
so that it calls `matchAndRewriteFixedRotationOrPhaseShift` or
`matchAndRewriteArbitraryRotation`.
3. Implement `matchAndRewriteArbitraryRotation` following the
mathematical formulas mentioned before.
4. Add tests for the general cases of `Rot` and `CRot`, and for the
special cases of merging arbitrary rotations.

Notice that, in order for the new code to compile, I have needed to
include a dependency to the MLIR Math dialect, as well as include guards
pointing to the Math header file.

**Benefits:**

Arbitary rotations can now be merged as part of Catalyst's Merge
Rotations pass.

**Possible Drawbacks:**

There are still a few TODOs in the code.
One of them, possibly the main one, has to do with the implementation of
the problematic scenarios for differentiability, which I haven't
addressed yet.

**Related GitHub Issues:**

#1220

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>