Set default diff_method to adjoint for lightning devices (#1961)
**Context:**
Currently parameter-shift is used by default for differentiation, when
the `diff_method` is not provided. This is the most universal method,
but not the most performant, especially on lightning devices which
support adjoint differentiation (for expval)

**Description of the Change:**
For lightning devices, set `diff_method` to `adjoint` when it is set to
`best`, which is the default argument.

Note: currently only a single expval is supported - this needs to be
fixed. See
[discussion](https://github.com/PennyLaneAI/catalyst/pull/1961#discussion_r2263890832)

**Benefits:**
Significantly faster differentiation performance on lightning devices.

**Possible Drawbacks:**

**Related GitHub Issues:**
https://github.com/PennyLaneAI/catalyst/issues/1171

[sc-74760]