Fix build system due to clash with new dialects docs build  (#1979)
**Context:** Some of our build commands, such as `make dialects`, were
failing since adding the build directives for our dialect documentation
in #1959. The cause of the issue was a new CMake variable
`CATALYST_DOCS_ONLY` that was clashing with the build system for other
targets.

**Description of the Change:** Build the dialect documentation in a
directory separate from the `build` directory used by the main dialect
build targets. This new directory is called `build-docs`.

---------

Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>