Improved `cond` support in `from_plxpr` (#1973)
**Context:**

We want to switch to using program capture as the frontend for catalyst,
so we need better coverage of features.

**Description of the Change:**

Allows `cond` to be translated when it is outside the qnode.

Allows `cond` to have elif branches.

Allows an empty false branch when the function has arguments.

Makes sure the predicates are always booleans.

A couple other smaller fixes.

**Benefits:**

Closer to using plxpr as the capture frontend.

**Possible Drawbacks:**

Still some open questions and tests marked xfail for now. 

**Related GitHub Issues:**
[sc-95436]