Enable end-to-end PCPhase support (#1987)
**Context:**
LQ and LGPU provide native support for PCPhase. To enable this in
Catalyst, we had to define a new custom MLIR gate op in the Quantum
dialect and extend both the frontend and runtime to lower PCPhase into a
new QIR instruction. This allows execution on devices with native
support and unlocks simulation of QSVT workloads w/ ~30 qubits on a
single GPU.

**Description of the Change:**

**Benefits:**
- End-to-end PCPhase support
- Simulate QSVT workflows with ~30 qubits backed by LGPU

**Possible Drawbacks:**
- We need a better representation in Catalyst to support generic PL
gates with hyperparameters.

**Related GitHub Issues:**
[sc-98149]