A new JAX primitive for quantum definitions is added to Catalyst (#1820)
**Context:**
A new JAX primitive for decomposition-rule definitions is added to
Catalyst.

This functionality is not supposed to be used by users, instead it's
used in the decomposition system to lower rules from PennyLane --> PLxPR
--> MLIR.


**Feature PRs:**
- [X] Added a new JAX primitive to capture and compile the decomposition
rule definitions to MLIR. `decomposition_rule` is the decorator
integrated with this primitive for development purposes.
- [X] Renamed `QregManager` to `QubitHandler` and extended the class to
manage
  converting PLxPR wire indices into Catalyst JAXPR qubits.
  This is especially useful for lowering decomposition rules that take
  in qubits as arguments.

[sc-95656]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>