Dynamic allocation core logic, runtime label resolving design (#2008)
**Context:**
The plan is to lower `qml.allocate` to a sort of "label generator
instruction", instead of lowering to `quantum.alloc` operations.
This is because `quantum.alloc` operations actually create new qubits on
the device, and that is very hard to work with the interpretation where
`qml.device(wires=X)` means `X` is the capacity. We would need to keep
the initial allocation size somewhere else, but that does not exist yet.

A strategy is to make all user wire labels in python to be just labels,
instead of interpreting them as the actual address of the qubit on the
device's qubit register array. Then `qml.allocate` can generate special
labels (e.g. negative labels) and work inside this label-to-index
resolution system itself.


**Description of the Change:**
This PR implements the infrastructure to map user labels to device qubit
indices.
Note that this PR does not touch `qml.allocate`. This is just infra.

When a user does `qml.something(wires=wire)`, this will be lowered to a
`quantum.extract [wire]` operation, which will then be lowered to
`__catalyst__rt__array_get_element_ptr_1d(wire)`.

In the current system, `__catalyst__rt__array_get_element_ptr_1d(wire)`
returns the qubit ID on the register array at index `wire`. This means
the value for `wire`, coming from user's `qml.thing(wires=wire)`, is
directly the qubit address.

We change this into the following, in order to interpret `wire` as just
a label:
1. The runtime device (`RTDevice`) would carry a map from the labels to
the actual indices into the register array. Indices start from 0. Every
time a new label is encountered, the next index is added into the map.
2. Logically, that was really the only change needed. However, a caveat
is that when user requests terminal measurements without wires, like
`qml.probs()`, they are implicitly requesting
`qml.probs(wires=[0,1,2,3,...])`, aka all wires on the device. However,
these wires are now labels, but just calling
`getQuantumDevicePtr()->Probs(view)` will ask the device to compute
these terminal measurements with the wire ID ordered by the allocated
qreg array, i.e. in index order. There needs to be a remapping from the
index-ordered results returned by the device back into the label-ordered
results the user wanted.