Fix result handling for one-shot (#2017)
**Context:**

The result of non-mcm cases are not process properly. This PR is going
to fix the result of non-mcm cases.

**Description of the Change:**

1. Fix normal result shape handling
2. Add support for CountsMP result shape handling
3. Add support for Shot vector result shape handling
4. Add support for snapshot result handling

**Test**

1. `test_mitigation.py` 
Lit test of `test_mitigation.py` needs to be updated. The reason for
this change is in the handling of the result, where the `is_valid`
parameter provided to gather_non_mcm has been changed from:
```python
is_valid = jnp.array([True] * len(out[i]))
```
to:
```python
is_valid = jnp.full((result.shape[0],), True)
```
Originally, this required passing `%c = stablehlo.constant dense<true> :
tensor<{shots}xi1>` to the one-shot wrapper. This modification
eliminates the need to pass that extra parameter. Instead, the logic is
handled within the body by using broadcast:
```mlir
%c = stablehlo.constant dense<true> : tensor<i1>
%1 = stablehlo.broadcast_in_dim %c, dims = [] : (tensor<i1>) -> tensor<{shots}xi1>
```

2. `test_mid_circuit_measurement.py`

`value_and_grad` now will be supported (fix the result handling), but
jax has an issue with the current version we use as it results in a
random memory error. https://github.com/tensorflow/tensorflow/pull/97681
will be available in the next jax release. we will re-enable this test
when the jax release is available and tested on our end.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-93649]

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>