Support array-backed registers (#2000)
**Context:** Currently, the `quantum.insert` op in Catalyst's quantum
dialect is effectively a no-op once it has been lowered to the LLVM
dialect (and therefore also in the runtime). For example, in the
following program,

```mlir
%0 = quantum.alloc( 1) : !quantum.reg
%1 = quantum.extract %0[ 0] : !quantum.reg -> !quantum.bit
%out_qubits = quantum.custom "<gate>"() %1 : !quantum.bit
%2 = quantum.insert %0[ 0], %out_qubits : !quantum.reg, !quantum.bit
```

the `insert` op does not actually "insert" the qubit represented by the
SSA value `%out_qubits` into the register (represented by value `%0`) at
position 0 at runtime. In other words, it does not manipulate the array
of qubit IDs represented by the `quantum.reg` type and update the qubit
ID at position 0 in the array. Instead, it simply replaces the uses of
the register returned by the `quantum.insert` op with the register given
as its input operand, as can be seen in the source code here:


https://github.com/PennyLaneAI/catalyst/blob/38064f849a17653ff176ce0a9d0478759a1d0222/mlir/lib/Quantum/Transforms/ConversionPatterns.cpp#L415-L421

It would be useful to have the ability to insert individually allocated
qubits, or qubits from an auxiliary register, into a specific position
in a given register (assuming the qubits at those locations have already
been individually deallocated). Indeed, this is an essential component
of nearly all MBQC workloads. We say that compilers with this ability
support *array-backed registers*.

**Description of the Change:** This PR contains two major additions:

1. The addition of a new runtime stub
    ```cpp
void __catalyst__rt__array_update_element_1d(QirArray *ptr, int64_t idx,
QUBIT *qubit)
    ```
which updates the value in the `QirArray` at position `idx` to be the
value of `qubit`.
2. The addition of the `InsertOpArrayBackedPattern` conversion pattern
that lowers `quantum.insert` ops to a call to this runtime stub during
conversion to the LLVM dialect.

The original `InsertOpPattern` conversion pattern has been renamed to
`InsertOpDefaultPattern`. **By default, Catalyst workloads will continue
to use the original conversion pattern.** A `use-array-backed-registers`
option has been added to the `convert-quantum-to-llvm` pass to opt in to
use the `InsertOpArrayBackedPattern` conversion pattern instead.

The MBQC pipeline, available via `catalyst.ftqc.mbqc_pipeline()`, has
also been updated to enable array-backed registers.

For example, this option can be enabled from the command line with,

```console
$ quantum-opt --pass-pipeline="builtin.module(convert-quantum-to-llvm{use-array-backed-registers=true})"
```

**Benefits:** Array-backed registers are essential to represent MBQC
workloads. This PR will enable ongoing and future FTQC development.

**Possible Drawbacks:** There is a risk in the runtime stub that we
overwrite the ID of an active qubit contained in the register (one that
has not been deallocated). Ideally we should be able to query whether a
qubit is active given its ID, but we do not have a general,
device-agnostic way to do this yet. For NullQubit, for example, we can
query its `qubit_manager` object to determine if a given qubit is
active, but not all devices implement a qubit manager.

[sc-97937]