Add constraints for one-shot and mechanism to fallback to the single-branch (#2019)
**Context:**

For some cases in one-shot we need to raise an error or fall back to a
single-branch

**Description of the Change:**

This change introduces error handling for the one-shot. For specific
usage that are not supported in this mode, this PR will now either raise
a error or automatically fall back to a single-branch. And this PR will
now default to the one-shot whenever `shots > 0`. The following
exceptions will trigger a fallback to the single-branch or raise an
error.

**Original unsupported measurements for one-shot will trigger to
fallback to single-branch:**
1. Native mid-circuit measurement mode does not support
2. qml.var(obs) cannot be returned when `mcm_method='one-shot'

**New exceptions after this PR:**
1. empty wires is not supported with dynamic wires in one-shot mode:
`qml.sample()` and `qml.counts()` will fall back to single-branch if no
constant number of wires is specified. This is necessary as JAX does not
support inserting a dynamic tensor to dynamic tensor (see issue #1949).
2. Transform measurements **from user** (e.g. measurement_from_counts,
measurement_from_sample) are incompatible with one-shot. The failure
occurs because of the timing at which these transformations are applied.
This is a framework-level timing issue, as other backends like `qrack`
can handle these correctly.

**Original exception but not raise when using one-shot**
Shot vector for non sample measurement is not supported, it will raise
an error.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-93649]

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>