The runtime `qubit_release_array` CAPI should release only the qubits in the array, not all qubits on the device (#1996)
**Context:**
The runtime `qubit_release_array` CAPI should release only the qubits in
the array, not all qubits on the device.
This is especially important for dynamic qubit allocation work. When
lowering `qml.deallocate`, the current `quantum.dealloc` op will be
lowered to the `qubit_release_array` to release qubits in the register,
but because the dynamically allocated qregisters are only temporary in
the program, we cannot just release all qubits on the device.

We were ok with this because there was always just one global qreg. But
now with dynamic allocation the asymmetry between batched allocation and
(the currently non-existing) batched release needs to be properly
handled.

**Description of the Change:**
- Add `ReleaseQubits(std::vector<QubitIdType>)` interface in
`QuantumDevice`, and remove `ReleaseAllQubits()`;
- `qubit_release_array` CAPI uses
`ReleaseQubits(std::vector<QubitIdType>)` instead of
`ReleaseAllQubits()`.
- Add guards for current devices that don't support dynamic
allocation:oqc, oqd, openqasm to check that there is only one
`ReleaseQubits()` call in all their programs.


**Benefits:**
- Symmetry between qubit allocation and deallocation in device api!
- Better support implementation of dynamic qubit allocation.

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>