[Decomposition] Update default fallback message following changes in PL (#2091)
**Context:**
Update the default fallback message and behaviour so that the system
always falls back to the legacy implementation when PL’s decomposition
graph fails to solve. This ensures consistency with PL `UserWarning`s
and allows us to continue compiling and executing the circuit
end-to-end.

**Description of the Change:**
- Capture warnings and raise a `UserWarning` in `from_plxpr` when graph
fails.
- Consistency with the behaviour of PL when the graph-decomposition
fails to find rules

**Benefits:**
- Clearer user warnings/errors (consistent with PL warnings/errors)
using `qml.transforms.decompose`

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-100399]