Bump CPL rc versions to 13 (#2093)
Context:
In preparation for the Catalyst v0.13.0-rc release candidate (RC), the
GitHub action must point to the corresponding RC branches of Catalyst,
Pennylane, and Lightning.

Description of the Change:
Update the Catalyst RC branch v0.12.0-rc -> v0.13.0-rc, as well as the
corresponding RC branches for Lightning and PennyLane from v0.42.0_rc to
v0.43.0_rc and v0.42.0-rc0 to v0.43.0-rc0.