Fixing incorrect outputs and crashes in dynamic one-shot when shots = 1 (#2090)
**Context:** Dynamic one-shot execution causes a crash or yields
incorrect output when we set `shots=1`. (See related issues #2087 and
#2088 for details). This issue was caused by an oversight when using the
function `jnp.squeeze`. When `shots=1`, the code removes the first axis
of the output array, which corresponds to the number of shots. This
interferes with subsequent calculations and leads to the crashes and
incorrect results as mentioned.

**Description of the Change:** 
This change modifies the output handling to correctly preserve the shots
dimension in the case when `shots=1`.

closes #2087 #2088

[[sc-100800]]
[[sc-100803]]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>