Modify RC CI to install RC branches from PL and Lightning (#2095)
**Context:**
PRs targeting `v0.13.0-rc` should be tested with the release candidate
branches of PL and Lightning as well.
We manually overwrite the CI to install these directly after `make
frontend` (which gets the version from `.dep_versions`).

Note that when merging rc branch back to main after the release, this PR
must be reverted for `.dep_versions` to take back control of the
versions.