Selective unitary decomposition fallback to fix IR blow-up (#2100)
Do not fallback to QubitUnitary unless operator provides no
decomposition. Previously, a QubitUnitary would be constructed for any
Controlled operator not natively supported by the device. While this can
have some advantages, e.g. better performance on simulators for gates up
to a certain size (which was never the intention of the fallback), it
can also cause issues like overly large constant matrices embedded into
the IR, and can also be unexpected to users since they might expect a
decomposition to elementary gates.

The proposed update should be more robust in that it still uses the
QubitUnitary fallback for gates that don't implement a decomposition,
but otherwise makes use of the provided one.

closes #2077