Fix a clang-tidy warning on main (#2112)
**Context:**
A clang-tidy warning on main branch is blocking benchmark runs (the reference runs). 

We do have `-Wall` turned on by default, but we might have slightly
different configs than the benchmark suite?
But anyway, this is the needed fix.

**Description of the Change:**
A class is instantiated with some fields missing initial values. 
We provide the empty init values.

**Benefits:**
Benchmark suite runs.