Fix weird consts issue in `from_plxpr` (#2128)
**Context:**

When updating tests for #2078 , I was getting weird failures on this
test:
https://github.com/PennyLaneAI/catalyst/blob/5eec7c770fc5fd1d5c5a9c03124db3967de8769d/frontend/test/pytest/test_gradient.py#L757

Basically what was happening was that the conversion of the target jaxpr
was reordering the consts. So we need to use the consts for the *new
jaxpr*, not the *input jaxpr*.

**Description of the Change:**

In all relevant translation steps, we pull the consts from the new
jaxpr, not the original inputs.

**Benefits:**

Get correct results without errors.

**Possible Drawbacks:**

I didn't test all the other control flow primitives that might have the
same issue, because I don't know how to trigger this is other cases.

**Related GitHub Issues:**

[sc-101609]