Issue #2152: Fix `catalyst.qjit` and `catalyst.CompileOptions` docs render (#2156)
**Context:**
`catalyst.qjit` and `catalyst.CompileOptions` docs are not rendered
properly in a couple of sections.

**Description of the Change:**
The fixes just consist of adding a blank line before the first element
of a sublist.
I've also taken the opportunity to do some minor changes on the
`catalyst.qjit` and `catalyst.CompileOptions` pages (mostly to be
consistent with most of the descriptions):
- Fix `catalyst.CompileOptions` doc render of `autograph_include`
parameter.
- Start argument descriptions with a capital letter.
- End argument descriptions  with a period.
- Use verb forms in plural (e.g., store or disable).
- Use `Default is <default value>` (instead of sometimes `Defaults to
<default value>`).

**Benefits:**
Proper doc rendering.

**Possible Drawbacks:**
None.

**Related GitHub Issues:**
#2152