Enable decomposition of MultiRZ at MLIR (#2160)
**Context:**
This PR refactors the implementation of `SignatureAnalyzer` in the
decompose-lowering pass extending the analyser and the pass to support
special operator definitions (e.g., MultiRZ) along with CustomOps as new
patterns.

**Description of the Change:**
- Add a `BaseSignatureAnalyzer` class that servers a base analyzer in
decompose-lowering
- Add MultiRZ decomposition patterns to the lowering pass
- Fix the issue with trying to capture and compile unutilized abstracted
Adjoint and Controlled decomposition rules from the graph. (We don't
need these rules from the graph as we handle these symbolic ops using
Catalyst AdjointOp and ControlledOp). Avoiding this saves compilation
time and failures for templated rules.
- Include MLIR and Python tests.


**Related Issues:**
[sc-102152]

---------

Co-authored-by: Hong-Sheng Zheng <mathan0203@gmail.com>