Issue #1621: Fix DataView iterator segfault with axis of size 0 (#2164)
**Context:**
The `DataView` iterators suffer an underflow of unsigned types that can
lead to a segfault when presented an axis of size 0, as discussed in
#1621 and
[here](https://github.com/PennyLaneAI/catalyst/pull/1598#issuecomment-2779178046).

**Description of the Change:**
This change removes the `var - 1` type arithmetic that lead to the
underflows, which in turn prevents the segfaults, by adjusting the
indexing done in the for loop.

**Benefits:**
No more underflow errors and segfaults, accurate `view.size()` and
`std::distance(view.begin(), view.end())` results in the case of an axis
of size 0.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
Closes #1621 

[sc-88502]

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>