[Capture] handle grad and jacobian (#2078)
**Context:**

Handles the `grad` primitive from pennylane and adds a lowering for it.

This requires PennyLane PR:
https://github.com/PennyLaneAI/pennylane/pull/8357

Once that merges in, we can retarget this branch against the newer
pennylane.

**Description of the Change:**

Instead of translating the pennylane grad prim to the catalyst one, we
just register a lowering for the grad primitive. We also have to
register a translation rule where we are translating the target jaxpr.

**Benefits:**

Can handle grad and jacobian calls.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-100560]