Auto detect xdsl passes (#2169)
**Context:**
We want to be able to use the xdsl interpreter to apply user passes
without the need to specify the xdsl plugin manually.

**Description of the Change:**
if a pass uses xdsl implementation, the transform sequence module and
the transform itself is marked by adding a and attribute to the
operation. this attribute is used later to detect wether we should use
the xdsl compiler or not.

**Benefits:**
No need to specify the xdsl plugin path. 

**Possible Drawbacks:**
An extra walk over the program is added to detect the passes. This Could
be avoided by keeping a global state when lowering the passes, but it
would lose the locality of this information in cases with multiple
qnodes that have different transform schedules.

**Related GitHub Issues:**
[sc-102894]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>