Update for use of a single transform primitive (#2177)
**Context:**

See https://github.com/PennyLaneAI/pennylane/pull/8576

Right now each transform has its own primitive, so we have to register
each of those primitives in a whole bunch of different locations. This
is hard to manage, and can lead to memory leaks.

So we are just going to deal with a single trnasform primitive with a
`transform` kwarg that holds onto the `TransformDispatcher` object.

**Description of the Change:**

Update to have a single registration to handle transforms.

**Benefits:**

Less work to extend to new transforms.

**Possible Drawbacks:**

Harder to customize the behaviour for one specific transform.

**Related GitHub Issues:**

[[sc-103382](https://app.shortcut.com/xanaduai/story/103382)]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>