Support for PauliRot and PauliMeasure in Catalyst (#2145)
**Context:**

There is currently no way for users to write PPRs and PPMs in the
Pennylane frontend and have it lowered to the QEC dialect directly. Note
that this feature is only supported when `qml.capture.enabled()` is
true.

**Description of the Change:**

`qml.PauliRot` and `qml.pauli_measure` will be converted into the
appropriate `qec.ppr` and `qec.ppm` operations in the MLIR, allowing us
to use the existing ppr/ppm compilation passes directly, as well as
writing custom decomposition rules in terms of them.

**Benefits:** 

This allows researchers to test PPR/PPM compilation passes easily by
writing circuits in terms of PPRs and PPMs and directly integrate with
the Catalyst's `ppm_compilation` pass (and sub-passes).

**Related GitHub PRs:**

[PennyLaneAI/pennylane#8544](https://github.com/PennyLaneAI/pennylane/pull/8544)

[sc-101578]
[sc-102095]

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>
Co-authored-by: Sengthai Heng <sengthai37@gmail.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>
Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>