Enforce order of state prep and basis state (#2239)
**Context:**
In the lightning devices, the state prep operations write to the entire
state vector.
This means if any quantum gates came before state prep, they will be
overwritten.

**Description of the Change:**
We force the position of state prep and basis state primitives to be the
same as how they appear in the user code.
In other words, we do not allow topological sorting to have the freedom
to change their locations.

**Benefits:**
Correct results of certain programs where tape-level transforms and
template decompositions somehow change the location of set state
primitives.

**Related GitHub Issues:** #2235 

[sc-104798]