merge_rotations PPR support (#2224)
**Context:**
With the addition of QEC and PPRs, passes need to be updated to
accommodate them.

**Description of the Change:**
Adds a merge rule for PPRs to the `merge_rotations` pass, which merges
PPRs of the form

$$PPR(\theta) \cdot PPR(\theta') \to PPR(\theta + \theta'). $$

The rule also simplifies the case 

$$PPR(\theta) \cdot PPR(-\theta) = I \to \textnormal{no instruction}$$

 and 

$$PPR(\pi - \theta) \cdot PPR(\theta) \to \textnormal{no instruction}.$$

Note: PPRs use a different angle convention than `qml.PauliRot`, so
$PPR(\pi) = I$.

**Benefits:**
Simplifies mlir output with regards to PPRs, adds expected functionality
to the `merge_rotations` pass.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
N/A

[sc-104292]
[sc-104293]

---------

Co-authored-by: Sengthai Heng <sengthai37@gmail.com>