Rename annotate-function pass and move it to the gradient dialect (#2241)
**Context:**
The `annotate-functions `pass marks functions that cannot be
automatically differentiated. It scans functions for operations
incompatible with automatic gradients (quantum measurements, certain
custom calls, and callbacks without custom derivatives) and annotates
those functions with a `invalidGradientOperation`. It also propagates
the annotation up the call graph, so if a function calls an annotated
function, it gets annotated too. Later, when verifiers for gradient
operations are run they will fail with a clear error message.
Currently this pass is in the quantum dialect and grouped with other
quantum compilation passes.

**Description of the Change:**
This PR improves the naming of the pass. Move it to the gradient dialect
and schedule within the `lower-gradients` compilation stage

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**