Fix a bug in `setup_dev_from_wheel` script (#2217)
**Context:**
`setup_dev_from_wheel` script aims to find the correct commit that
corresponds to the latest PyPi release, however it checks out the parent
commit of the nightly version bump commit which would have a lower
version and hence fails to find the correct commit.

**Description of the Change:**
changes the script to checkout the correct commit.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-104391]