Use `pass_name` from pennylane transform (#2149)
**Context:**

Trying to improve our integration between pennylane transforms and
catalyst passes.

See #2206 for where we are able to update all built-in catalyst passes
to be transforms instead. That PR helps validate we can make this change
with no breaking behaviour.

Going forward, we can setup a plan to get rid of the catalyst-specific
pass handling infrastructure.

Also note that this PR targets the PennyLane branch, so we can confirm
we won't need anymore changes there to get catalyst working correctly
before approving and merging.

**Description of the Change:**

Uses the `transform.pass_name` property as the higher priority source of
the `pass_name`.

Updates the `pass_pipeline` from the `quantum_kernel_p` primitive to be
a tuple of `TransformContainer` objects instead of `Pass`, but leaves in
the logic to handle `tuple[Pass,...]` for safer backwards compatibility.

Also makes using transforms with pass names by splitting a `QNode`'s
transform program into "things at the start of the program that are tape
transforms" and "things at the end of the program that are MLIR passes".
This will allow us to get rid of `PassPipelineWrapper` use for both
systems and soften the change from old frontend to new frontend.

**Benefits:**

Unified handling of transforms across pennylane and catalyst. 

**Possible Drawbacks:**
 


**Related GitHub Issues:**

Depends on https://github.com/PennyLaneAI/pennylane/pull/8539
[sc-103775]

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>