Remove the local transform dialect (#2261)
**Context:**
The local transform dialect was created in
https://github.com/PennyLaneAI/pennylane/pull/7956 because xdsl was not
up to date with Catalyst's LLVM version which updated the transform
dialect to accept dictionaries as options. Since xDSL has now updated
their LLVM version, we can remove the local dialect.
 
**Description of the Change:**
Remove the local shim for the transform dialect and use the xdsl
upstream one.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**